neural-compression/requirements.txt
torch
compressai
Pillow
numpy
scikit-image
matplotlib
tqdm
opencv-contrib-python>=4.5.0
networkx>=2.6.0
scipy>=1.7.0
pytest
